+++ lib/config/config_EXAMPLE_pcomm.rb
@@ -7,5 +7,6 @@ ButlerMainframe.configure do |config|
   config.session_path   = '"C:/Program Files (x86)/IBM/Personal Communications/pcsws.exe" "C:/Users/YOUR_USER/AppData/Roaming/IBM/Personal Communications/host3270.ws" /Q /H /S=A'
   config.session_tag    = 'A'
   config.timeout        = 6000
+ #config.env            = 'production'
 end