+++ lib/core/configuration.rb
@@ -2,7 +2,7 @@ module ButlerMainframe
   class Configuration
     attr_writer :allow_sign_up
 
-    attr_accessor :language, :host_gateway, :browser_path, :session_url, :session_path, :session_tag, :timeout
+    attr_accessor :language, :host_gateway, :browser_path, :session_url, :session_path, :session_tag, :timeout, :env
 
     def initialize
       @language       = :en
@@ -12,6 +12,7 @@ module ButlerMainframe
       @session_path   = ""
       @session_tag    = nil
       @timeout        = 1000
+      @env            = 'development'
     end
 
   end