+++ lib/generators/rails/butler_mainframe_test/butler_mainframe_test_generator.rb
@@ -1,0 +1,5 @@
+class Rails::ButlerMainframeTestGenerator < Rails::Generators::NamedBase
+  source_root File.expand_path('../templates', __FILE__)
+
+  hook_for :test_framework
+end